requirements.txt
py-clob-client>=0.0.1
python-dotenv>=1.0.0
websockets>=12.0
requests>=2.31.0
aiohttp>=3.9.0
